requirements.txt
fastapi 
uvicorn 
gitpython 
transformers 
torch 
networkx 
coverage 
requests